// src/models/models.ts
export interface InterpretedPrompt {
    applicableGoalKeys: string[];
    prioritizedNutrients: (keyof NutritionalInfo)[];
    foodsToInclude: string[];
    foodsToAvoid: string[];
    explanation: string;
    originalPrompt: string;
    isBudgetFriendly?: boolean;
}
// ...
export interface FoodItem {
    food: Food; // Objeto completo do alimento (standard_name, category, etc.)
    quantity: number; // Quantidade sugerida
    isSubstituted?: boolean; // Se foi uma substituição
    originalFood?: Food; // Alimento original (se substituído)
    orderItemId: string;
    // ✨ NOVO CAMPO ✨
    explanationInDiet?: string; // Ex: 'O Salmão é rico em Ômega-3, essencial para o seu objetivo de saúde cardiovascular.'
    originalQuantity?: number;

}

export interface Food {
    id: string;
    standard_name: string;
    synonyms: string[];
    category: FoodCategory;
    tags: FoodTag[];
    processing: ProcessingType;
    seasonality: string[];
    glycemic_index?: number;
    nutritional_info_per_100g: NutritionalInfo;
    estimatedPrice: number;
    variableWeight?: boolean;
    quantity: number;
    default_unit: string;
    weight_per_unit_in_g?: string;
    imageUrl?: string;
    pickingInstructions?: string;
    max_weekly_g_per_person?: number;
}

export type FoodCategory = keyof typeof categoryDictionary;
export type DefaultUnit = keyof typeof defaultUnitDictionary;
export type ProcessingType = keyof typeof processingDictionary;
export type FoodTag = keyof typeof tagDictionary;

export interface NutritionalInfo {
    energy: number; carbohydrates: number; sugars: number; fiber: number;
    starch: number; proteins: number; total_fat: number; saturated_fat: number;
    trans_fat: number; monounsaturated_fat: number; polyunsaturated_fat: number;
    cholesterol: number; water: number; vitamin_a: number; vitamin_b1: number;
    vitamin_b2: number; vitamin_b3: number; vitamin_b5: number; vitamin_b6: number;
    vitamin_b7: number; vitamin_b9: number; vitamin_b12: number; vitamin_c: number;
    vitamin_d: number; vitamin_e: number; vitamin_k: number; calcium: number;
    iron: number; magnesium: number; phosphorus: number; potassium: number;
    zinc: number; sodium: number; copper: number; manganese: number;
    selenium: number; iodine: number; chromium: number; molybdenum: number;
    silicon: number; vanadium: number; omega_3: number; omega_6: number;
}
// ...
export const categoryDictionary: { [key: string]: string } = {
    cereal: "cereal", legume: "legume", fruit: "fruit", vegetable: "vegetable",
    animal_protein: "animal_protein", dairy: "dairy", nut: "nut", seed: "seed",
    fat: "fat", sweetener: "sweetener", beverage: "beverage", coffee_tea: "coffee_tea",
    condiment: "condiment", spice: "spice", sauce: "sauce", bakery: "bakery",
    dessert: "dessert", canned_goods: "canned_goods",
};

export const defaultUnitDictionary: { [key: string]: string } = {
    g: "g", ml: "ml", unit: "unit(s)", kg: "kg", l: "l",
};

export const processingDictionary: { [key: string]: string } = {
    unprocessed: "unprocessed", minimally_processed: "minimally_processed",
    processed: "processed", ultra_processed: "ultra_processed",
};

export const tagDictionary: { [key: string]: string } = {
    vegan: "vegan", vegetarian: "vegetarian", pescatarian: "pescatarian", gluten_free: "gluten_free",
    dairy_free: "dairy_free", low_sugar: "low_sugar", high_fiber: "high_fiber", low_fat: "low_fat",
    high_protein: "high_protein", low_sodium: "low_sodium", low_calorie: "low_calorie", low_carbs: "low_carbs",
    high_iron: "high_iron", high_calcium: "high_calcium", antioxidants: "antioxidants", keto: "keto",
    paleo: "paleo", organic: "organic", non_gmo: "non_gmo", whole_grain: "whole_grain",
    fermented: "fermented", nut_free: "nut_free", soy_free: "soy_free", refined: "refined",
    processed: "processed", ultra_processed: "ultra_processed",
};









export interface Coordinates {
    lat: number;
    lon: number; // ou lng, como preferir
}

export interface Address {
    id: string;
    street: string;
    number: string;
    complement?: string;
    neighborhood: string;
    city: string;
    state: string;
    zipCode: string;
    isDefault?: boolean;
    coordinates?: Coordinates;
}

export interface HealthProfile {
    healthConditions: string[];
    allergies: string[];
    currentMedications: string[];
    dietaryRestrictions: string[];
    favoriteFoods?: string[];
    activityLevel: string;
    isActivityLevelDetailed: boolean;
    activityLevelDetail: string;
    bodyFatLevel: string; // 1. Magro(a) | 2. Atlético(a) | 3. Em Forma | 4. Acima do Peso;
    bodyFatPercentage: number;
    metabolism?: string;
    sex: string;
    height: number;
    weight: number;
    dateOfBirth: string;
}